refactor: 部分代码结构调整，提升可维护性
